Package completions for aliased imports
