+++ Cargo.toml
@@ -18,7 +18,7 @@ path = "src/lib.rs"
 serde_json = "1.0"
 serde_repr = "0.1"
 serde = { version = "1.0.106", features = ["derive"] }
-flux = { git = "https://github.com/influxdata/flux", rev = "840a526" }
+flux = { git = "https://github.com/influxdata/flux", rev = "85bd204" }
 url = "2.1.1"
 lazy_static = "1.4.0"
 wasm-bindgen = "0.2.62"
