+++ src/handlers/mod.rs
@@ -51,10 +51,7 @@ pub fn create_diagnostics(
     let errors = check::check(walker);
     let diagnostics = map_errors_to_diagnostics(errors);
 
-    match create_diagnostics_notification(uri, diagnostics) {
-        Ok(msg) => Ok(msg),
-        Err(e) => Err(format!("Failed to create diagnostic: {}", e)),
-    }
+    Ok(create_diagnostics_notification(uri, diagnostics))
 }
 
 #[derive(Default, Clone)]
